+++ internal/iac_catalog/blueprint_config/blueprint_config_data_source_test.go
@@ -752,29 +752,6 @@ func TestBlueprintConfigConditionalsReading(t *testing.T) {
 			"Cloudfront": `{
 				"id": "clbnr5y2019144hyi1xc0yhex",
 				"variables": [
-					{
-						"id": "Cloudfront.comment",
-						"conditionals": [
-						  {
-							"source": "generic.variable.environment",
-							"condition": "nonprod",
-							"options": [
-							  {
-								"label": "nonprod",
-								"fieldId": "",
-								"value": "fake string",
-								"checked": false
-							  },
-							  {
-								"label": "sandbox",
-								"fieldId": "",
-								"value": "abc123fgh",
-								"checked": true
-							  }
-							]
-						  }
-						]
-					  },
 					  {
 						"id": "Cloudfront.dummy",
 						"conditionals": [
@@ -799,9 +776,8 @@ func TestBlueprintConfigConditionalsReading(t *testing.T) {
 	}
 	assert.Equal(t, 1, len(blueprintConfig.Children))
 
-	assert.Equal(t, 2, len(blueprintConfig.Children["Cloudfront"].Variables))
-	assert.Nil(t, blueprintConfig.Children["Cloudfront"].Variables[0].Conditionals[0].Value)
-	assert.Equal(t, "dummy-static-value", *blueprintConfig.Children["Cloudfront"].Variables[1].Conditionals[0].Value)
+	assert.Equal(t, 1, len(blueprintConfig.Children["Cloudfront"].Variables))
+	assert.Equal(t, "dummy-static-value", blueprintConfig.Children["Cloudfront"].Variables[0].Conditionals[0].Value)
 
 	jsonBlueprintConfig, _ := utils.ToJsonString(blueprintConfig)
 	fmt.Printf("blueprint config: [%v]\n", jsonBlueprintConfig)